Make undefined data references return false
This commit makes data references that are undefined return a false
value.

If a data reference is undefined the module now considers this to
evaluate to false, rather than throw an error.

A use-case for this would be when implementing logic to handle rules
that are not mandatory, which would need to be evaluated if they are
part of the data object, but not error out if they are not.